+++ ErsatzTV.FFmpeg/Capabilities/IHardwareCapabilitiesFactory.cs
@@ -1,3 +1,5 @@
+using ErsatzTV.FFmpeg.Capabilities.Qsv;
+
 namespace ErsatzTV.FFmpeg.Capabilities;
 
 public interface IHardwareCapabilitiesFactory
@@ -12,6 +14,8 @@ public interface IHardwareCapabilitiesFactory
         Option<string> vaapiDevice);
 
     Task<string> GetNvidiaOutput(string ffmpegPath);
+
+    Task<QsvOutput> GetQsvOutput(string ffmpegPath, Option<string> qsvDevice);
 
     Task<Option<string>> GetVaapiOutput(Option<string> vaapiDriver, string vaapiDevice);
 }